requirements.txt
datasets>=2.18.0
huggingface-hub>=0.21.0
pillow>=10.0.0
tqdm>=4.66.0
requests>=2.31.0
pip install numpy matplotlib scikit-learn
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
fastapi
uvicorn
pip install python-multipart
c2pa-python